The ninth grip - you can turn things now, which was the third word all along
> *"unfortunately there was no way to reposition, resize, or rotate it on the
> screen. Can I please please please have that too?"*

Reposition landed this evening and resize with it. **Rotate is the third word in
that sentence** and had no affordance at all: `handles.md` H2 says the standard
set is eight resize grips, a body and a rotation handle offset outside the top
edge, and this shell had eight and a body.

Select something and there is a circle above the box on a short stem. Drag it
and the selection turns about its own centre, previewed as you go. Hold Shift
and it steps in 15 degrees - which is what PowerPoint, Illustrator, Inkscape and
Figma all use, and it is 15 rather than 10 because 15 divides 90 AND 360, so the
four right angles and the four diagonals are all reachable.

★★ A ROTATION IS NOT A RESIZE WITH DIFFERENT ARITHMETIC, and the module boundary
follows from one row of the table:

  |                                  | resize            | rotate          |
  | measured from                    | opposite corner   | the CENTRE      |
  | what the drag reads              | a displacement    | a bearing       |
  | what the pointer's DISTANCE means| everything        | NOTHING         |

The third row is the one that decides it. An operator swinging a long arc for
precision is doing exactly what the gesture invites, and a build that let the
radius in would shrink the object while they did it. There is a test whose only
job is that property.

★★ THREE LINKS WOULD EACH HAVE PRODUCED A WORKING GESTURE AIMED AT THE WRONG
VERB, and every one of them looks deliberate from a chair:

- `Grip::is_resize` was `self != Self::Move`. Left alone, the rotate handle
  becomes the NINTH RESIZE GRIP - an object that gets bigger when you meant to
  turn it. It is enumerated now rather than negated, so a tenth affordance has
  to be classified rather than defaulting into the resize family.
- The routing arm below it read `(None, Some(_)) => DragKind::Move`. Left alone,
  a press on the handle MOVES the selection. Matched by name now.
- `needs_targets` had to grow `Rotate`, which is the FOURTH time that list has
  been the defect - `Resize`, `Handle` and `DimensionVertex` each shipped
  needing the decomposition and not asking for it.

★ THE NEGATION HAPPENS EXACTLY ONCE. Screen y is down, so a clockwise drag reads
as +90 from `atan2`; PDF user space is y-up and `Matrix::rotate` turns
anticlockwise in it. The single `-` lives in `rotating::drag` at the page
crossing, and the ghost is drawn from the UN-negated angle in screen space - so
the only thing that can differ between preview and commit is that sign, which is
a unit test rather than a hope. Getting it wrong gives a perfectly good rotation
the wrong way round, which nobody who did not watch the pointer would call a
defect.

★ And the drag past 180 degrees turns the SHORT way. Without normalising into
(-pi, pi] a pointer moving smoothly through the ray behind the centre makes the
object spin a whole turn in one frame - it looks like a physics bug and it is an
arithmetic one.

The ghost draws a QUADRILATERAL, not a rotated rect. Rotating the bounding box
would show a shape that grew as you turned it - a preview of something the
release does not do.

DRIVEN, live, on his own drawing:

    rotate-commit deg=-90.22 px=244.22 py=364.54 objects=1 constrained=0
    ★★ and it turned the right way
    ★★ the rotation reached the engine through `transform_objects`

The check builds every point with `declared_at` rather than assembling a
`ScreenPoint` from two integers - which it was the first place in this harness
to be tempted to do, and `ScreenPoint`'s single constructor is the discipline
that stopped it. A fraction is not a distance until you say of what, which is
why both extents of the box are read.

16/16 gates. 1,575 tests.

+++ tools/ui-verify/src/checks/insert_image.rs
@@ -520,13 +520,14 @@ fn drive(ctx: &CheckContext, report: &mut CheckReport) -> Result<Option<String>>
     //
     // Unreadable geometry leaves the floor at 1, which is the old behaviour's
     // honest core: what must not happen is ZERO.
-    let page_pts = crate::fixture::page_geometry(&pdf).map_or(0.0, |g| f64::from(g.width_pt));
+    let page_pts = crate::fixture::page_geometry(&pdf).map_or(0.0, |g| g.width_pt);
     let scale = if page_pts > 0.0 {
         f64::from(rect.w) / page_pts
     } else {
         0.0
     };
-    let expected = (f64::from(FIXTURE_W) * scale).max(1.0) * (f64::from(FIXTURE_H) * scale).max(1.0);
+    let expected =
+        (f64::from(FIXTURE_W) * scale).max(1.0) * (f64::from(FIXTURE_H) * scale).max(1.0);
     #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
     let floor = (expected / 4.0).max(1.0) as u64;
     if total > 0 && differing < floor {
